seqlib: stable-sort: fix bogus optimization's in #'stable-sort.
This patch fixes bug #292, where #'stable-sort weren't sorting in
place strings, when provided with predicate #'char-greaterp. This was
due to optimization, where if sequence was of type string or
vector-bit function #'sort was called (which invoked #'quicksort),
while for the rest cases merge-sort is used.

Optimization is reduced to use quicksort only on vector-bit sequences.

See: http://sourceforge.net/p/ecls/bugs/292/