format: format-fixed-aux: trailing zero is obligatory.
Trailing zero is obligatory unless d is explicitly set to 0. Consider
following example:

(format nil "~0f" 3.0) ; 3.0

If trailing zero is optional, then 3. would be printed and this would
be an integer (not float). If d is explicitly set to 0 we assume, that
programmer knows what he's doing and tries to convert float to
integer.

Signed-off-by: Daniel Kochmański <dkochmanski@turtle-solutions.eu>