random-state: allow initialization from the array
Array used for initialization is an array of the final values of
internal vector used to generate next randoms which should be a
product of initialization from a random seed.

This ability is required by CLHS to be able to read back the printed
random state. To print readable representation of random state
*print-readably* should be bound to T.

Signed-off-by: Daniel Kochmański <daniel@turtleware.eu>