typep: handle * type
This is the same as T according to '2.4.3 Type Specifiers' in the
spec. Fixes #240.

"If a type specifier is a list, the car of the list is a symbol,
and the rest of the list is subsidiary type information. Such a type
specifier is called a compound type specifier. Except as explicitly
stated otherwise, the subsidiary items can be unspecified. The
unspecified subsidiary items are indicated by writing *. For example, to
completely specify a vector, the type of the elements and the length of
the vector must be present."