buildsystem: split config.h into config.h and config-internal.h
Part of config.h was removed during the installation, we split config.h
so that part is in a separate file.

Additionally we add ECL_BUILD define during ECL build phase, so we can
distinguish whenever we build ECL or use it normally (because ecl.h has
to include config-internal.h at the build time).