unixfsys.d: fix symlink buffer grow
We were verifying `written == size' to check if there is something more
to be read, but previously we did `size+=256', so this test was always a
failure.

Additionally we fix a bug (not reported anywhere) for symlinks which
have 128+256n characters – '/' for directories and '\0' in the end
wouldn't fit in this corner cases.

Related to #295.