unixfsys: remove coerce_to_posix_filename
This was a workaround around bugs in cygwin and mingw (which are already
fixed), when the directory name ended with '/'. This fix introduced
extra bug with probe-file called on the root directory (it returned nil
on all platforms).

Original comment:

This converts a pathname designator into a namestring, with the
particularity that directories do not end with a slash '/', because this
is not supported on all POSIX platforms (most notably Windows).

Bug reference:

https://sourceforge.net/p/mingw-w64/patches/34/

Fixes #351.