Check for a symbol before attempting to bind it in PROGV
PROGV was attempting to bind whatever was in its variable list
without checking its type.  Using either the C-compiler or bytecode
compiler/interpreter, the following example would lead to a segfault
on my OpenBSD and Linux boxes:

> (defun foo () (progv (list 3) (list 3)))
FOO

> (foo)
Condition of type: SEGMENTATION-VIOLATION
[...]

Now give an error when attempting to bind something that is not a
symbol in PROGV (in both the C-compiler and bytecode compiler and
interpreter).