No longer allow LET/LET* to bind constants in bytecode compiler/interpreter
LET/LET* were allowed to lexically and dynamically bind constants in
the bytecode compiler and interpreter:

> (let ((pi 3)) pi)
3

> (progn (defconstant +c+ 'foo) (let ((+c+ 'bar)) +c+))
BAR

> (flet ((hello () (format t "hi")))
    (let ((t nil))
      (declare (special t))
      ; Oops, now this returns a string
      (hello)))
"hi"

Plus plenty of other ways to wreak havoc on unsuspecting code.

CLHS says the behavior is undefined when attempting to bind or assign
constant variables (CLHS 3.1.2.1.1.3 and the entry for defconstant).
(Well, CLHS 3.4.1 explicitly says that constant variables cannot be
used for variables in lambda lists.)

The C-compiler gives errors for these sorts of things, and the bytecode
compiler and interpreter gives errors when attempting to bind or assign
constant variables in lambda expressions, SETQ and various other forms.
So the behavior above in LET is inconsistent with both the C-compiler
and other parts of the bytecode compiler and interpreter.

Now give an error when attempting to bind a constant variable in
LET/LET* in the bytecode compiler and interpreter.

This also changes the behavior of PROG/PROG* and DESTRUCTURING-BIND so
that they give errors when attempting to bind constants as well.