No longer allow M-V-B to bind constants in bytecode compiler/interpreter
M-V-B was allowed to lexically and dynamically bind constants in the
bytecode compiler and interpreter:

> (multiple-value-bind (pi rem) (truncate pi) pi)
3

CLHS says the behavior is undefined when attempting to bind or assign
constant variables (CLHS 3.1.2.1.1.3 and the entry for defconstant).

The C-compiler gives errors for these sorts of things, and the bytecode
compiler and interpreter gives errors when attempting to bind or assign
constant variables in lambda expressions, LET, SETQ and various other
binding/assignment forms.  So the behavior above in M-V-B is
inconsistent with the C-compiler and other parts of the bytecode
compiler and interpreter.

Now give an error when attempting to bind a constant variable in M-V-B
in the bytecode compiler and interpreter.