bytecmp: tagbody: ensure ECL_NIL->ECL_NIL_SYMBOL conversion
Issue comes from the fact that list and symbol types are not
disjoint. Fixes #475.

My personal opinion (to put some rambling in a commit message) is that
having NIL being so many things is a blatant mistake and shouldn't be
picked up for Common Lisp (though many programs would probably break
and it wouldn't be so Common then).