parse_number: float: do not truncate result
Until now we did parse the number and the exponent and then we did
computation on said numbers (what implied ratio->float conversion
which right now truncates). Since we do depend on c99 already we use
standard functions strtof, strtod and strtold so we construct float
out of the bat. That prevents truncating floats on read.