tests: fix output-streams test from the run-program testsuite
The test was failing because it was detecting an output of

"

Hello stdoutHello stderr"

instead of the expected

"
Hello stdout
Hello stderr"

The reason for this is that the buffers of the
streams *standard-output* and *error-output* may be distinct, even if
the ouput send to both streams ends up in the same place. Thus, we
need to explicitely force-output in print-test. The error manifested
itself only in builds without thread support, because otherwise ECL
uses write instead of fwrite, avoiding any buffering.