Always close stream in with-output-to-string & cosmetic changes
Fix #576, Related to !197, 72560efa5a4300e6

with-output-to-string is required to close the output stream that it provides
for the extent of the body forms [1]. The current definition does not do that.

This change wraps the body forms in unwind-protect clauses to ensure the stream
is always closed on exit. Because declarations cannot appear at the beginning of
progn forms, any potential declarations are extracted from the body forms and
moved to the beginning of the surrounding let form's body.

element-type is no longer bound to a gensym, but evaluated inside the let body.

The uppercased names are downcased for a more coherent appearance.

[1]: http://www.lispworks.com/documentation/HyperSpec/Body/m_w_out_.htm