Add a regression test for the bug described in #576
Merge a test for with-output-to-string with the one for
with-input-input-from-string to ensure both close their streams.

Remove check for stream-var being a stream outside of w-i-f-s &
w-o-t-s. According to the specification, the streams' extent ends with the
respective providing form. If the stream was indeed not acccessible anymore, the
test would not pass. In that case open-stream-p should signal a type-error,
causing the test to crash. However in ECL we can assume that the stream is still
intact.